Revert "Metrics output plugin"
