+++ splunk_eventgen/eventgen_core.py
@@ -265,7 +265,6 @@ class EventGenerator(object):
         eventgen_metrics_logger_path = os.path.join(log_path, 'eventgen-metrics.log')
         eventgen_error_logger_path = os.path.join(log_path, 'eventgen-errors.log')
         eventgen_server_logger_path = os.path.join(log_path, 'eventgen-server.log')
-        eventgen_httpevent_logger_path = os.path.join(log_path, 'eventgen-httpevent.log')
         if not config:
             log_format = '%(asctime)s %(name)-15s %(levelname)-8s %(processName)-10s %(message)s'
             date_format = '%Y-%m-%d %H:%M:%S'
@@ -304,11 +303,6 @@ class EventGenerator(object):
             server_file_handler.setFormatter(json_formatter)
             server_file_handler.setLevel(logging.INFO)
 
-            httpevent_file_handler = logging.handlers.RotatingFileHandler(eventgen_httpevent_logger_path, maxBytes=2500000,
-                                                                       backupCount=10)
-            httpevent_file_handler.setFormatter(detailed_formatter)
-            httpevent_file_handler.setLevel(logging.INFO)
-
             # Configure eventgen logger
             logger = logging.getLogger('eventgen')
             logger.setLevel(self.args.verbosity or logging.ERROR)
@@ -342,12 +336,6 @@ class EventGenerator(object):
             logger.handlers = []
             logger.addHandler(server_file_handler)
             logger.addHandler(console_handler)
-
-            logger = logging.getLogger('eventgen_httpeventout')
-            logger.setLevel(logging.INFO)
-            logger.propagate = False
-            logger.handlers = []
-            logger.addHandler(httpevent_file_handler)
         else:
             self.logger_config = config
             logging.config.dictConfig(self.logger_config)
